Move framework/binary/bitstream* to core/data/binary
