+++ core/data/binary/bitstream_test.go
